rapid_relief_app/requirements.txt
snowflake-snowpark-python
streamlit
pandas
plotly
tabulate
requests
gTTS
toml
